Removed the /internal/ directory in Keyvault Keys and move headers to private that don't need to be public.
