+++ sdk/keyvault/azure-security-keyvault-keys/inc/azure/keyvault/keys/cryptography/cryptography_client.hpp
@@ -14,12 +14,12 @@
 #include <azure/core/io/body_stream.hpp>
 
 #include "azure/keyvault/keys/cryptography/cryptography_client_options.hpp"
+#include "azure/keyvault/keys/cryptography/cryptography_provider.hpp"
 #include "azure/keyvault/keys/cryptography/encrypt_parameters.hpp"
 #include "azure/keyvault/keys/cryptography/encrypt_result.hpp"
+#include "azure/keyvault/keys/cryptography/remote_cryptography_client.hpp"
 #include "azure/keyvault/keys/cryptography/sign_result.hpp"
 #include "azure/keyvault/keys/cryptography/wrap_result.hpp"
-#include "azure/keyvault/keys/internal/cryptography/cryptography_provider.hpp"
-#include "azure/keyvault/keys/internal/cryptography/remote_cryptography_client.hpp"
 
 #include <memory>
 #include <string>
@@ -51,18 +51,6 @@ namespace Azure {
         bool forceRemote);
 
     void Initialize(std::string const& operation, Azure::Core::Context const& context);
-
-    /**
-     * @brief Provides a #CryptographyProvider that performs operations in the Key Vault Keys
-     * Server.
-     *
-     * @return A cryptographic client to perform operations on the server.
-     */
-    std::shared_ptr<Azure::Security::KeyVault::Keys::Cryptography::_detail::CryptographyProvider>
-    RemoteClient() const
-    {
-      return m_remoteProvider;
-    }
 
     /**
      * @brief Gets whether this #CryptographyClient runs only local operations.
