+++ sdk/keyvault/azure-security-keyvault-keys/inc/azure/keyvault/keys/cryptography/remote_cryptography_client.hpp
@@ -17,7 +17,7 @@
 #include "azure/keyvault/keys/cryptography/cryptography_client_options.hpp"
 #include "azure/keyvault/keys/cryptography/encrypt_parameters.hpp"
 #include "azure/keyvault/keys/cryptography/encrypt_result.hpp"
-#include "azure/keyvault/keys/internal/cryptography/cryptography_provider.hpp"
+#include "azure/keyvault/keys/cryptography/cryptography_provider.hpp"
 #include "azure/keyvault/keys/key_vault_key.hpp"
 
 #include <memory>
