+++ sdk/keyvault/azure-security-keyvault-keys/src/cryptography/cryptography_client.cpp
@@ -9,8 +9,11 @@
 #include <azure/core/http/policies/policy.hpp>
 
 #include "azure/keyvault/keys/cryptography/cryptography_client.hpp"
-#include "azure/keyvault/keys/internal/cryptography/local_cryptography_provider_factory.hpp"
+#include "azure/keyvault/keys/cryptography/cryptography_provider.hpp"
+#include "azure/keyvault/keys/cryptography/remote_cryptography_client.hpp"
 #include "azure/keyvault/keys/key_operation.hpp"
+
+#include "../private/local_cryptography_provider_factory.hpp"
 
 #include <memory>
 #include <string>
