+++ sdk/keyvault/azure-security-keyvault-keys/src/cryptography/remote_cryptography_client.cpp
@@ -11,7 +11,8 @@
 #include "../private/key_sign_parameters.hpp"
 #include "../private/key_verify_parameters.hpp"
 #include "../private/key_wrap_parameters.hpp"
-#include "azure/keyvault/keys/internal/cryptography/remote_cryptography_client.hpp"
+
+#include "azure/keyvault/keys/cryptography/remote_cryptography_client.hpp"
 
 #include <memory>
 #include <string>
